+++ HISTORY.md
@@ -1,3 +1,9 @@
+## 1.0.2 (2012-12-??)
+
+* Replaced JUnit with TestNG
+* Updated Jackson dependency from 2.1.0 to 2.1.1
+* Added support for recurring e-mail receivers
+
 ## 1.0.1 (2012-11-18)
 
 * Billomat JSON-API now supports multiple e-mail recipients